refactor: coerce numeric literals at the parser, not at every consumer
Follow-up to enabling pyrefly: `unnecessary-type-conversion` flagged seven
`int()`/`float()` calls. Re-checking each against runtime behaviour rather
than the annotations showed the rule was mostly right and the annotations
were wrong.

The grammars stored ints for integer literals in fields declared float:
`NumberNode.value: float` held `int` for "3", and where-parser `Comparison.
value: float | str` held `int` for "snapshot > 0". Because int is assignable
to float (PEP 484 numeric tower) the type checker cannot see the difference,
so the inaccuracy was invisible and every consumer re-coerced defensively.

Coerce once at the parse action instead. That makes the declared types true
and the downstream calls genuinely redundant:

- lowering: `ir.Const(float(node.value))` -> `ir.Const(node.value)`, and the
  same for `_lower_bound` (pydantic already coerces BoundsDef to float).
- executor: `repr(float(p.value))` -> `repr(p.value)` for Cmp / DimCmp.
- builder: `_shift_amount` and the roll/shift helpers take `float`, which is
  what they always received — `shift(x, t=1.0)` is legal YAML. The
  `int(n) != n` integrality check stays and now reads as a real conversion.

`_lit` keeps `repr(float(v))` behind a one-line
`# pyrefly: ignore[unnecessary-type-conversion]`: its parameter is `float`,
so `_lit(0)` type-checks and would emit '0', which duckdb reads as INTEGER
rather than DOUBLE. All three ir.Const construction sites are float-pure
today; this keeps that from being load-bearing.

With the annotations honest, `unnecessary-type-conversion` is promoted to
error alongside the rest — one precise opt-out instead of the rule off.

Verified the LP text is byte-identical to main for examples/storage.yaml
(integer bounds, roll(), and integer where-comparisons), so no SQL literal
changed from DOUBLE to INTEGER or back.

+++ linopy_yaml/builder.py
@@ -352,7 +352,7 @@ def _helper_group_sum(array: Any, mapping: Any, *, into: str) -> Any:
     raise TypeError(msg)
 
 
-def _shift_amount(helper: str, kwargs: dict[str, int]) -> tuple[str, int]:
+def _shift_amount(helper: str, kwargs: dict[str, float]) -> tuple[str, int]:
     """Unpack and check the shared ``<dim>=<n>`` signature of roll/shift."""
     if len(kwargs) != 1:
         msg = f'{helper}() expects exactly one keyword argument (dim=n), got {len(kwargs)}: {kwargs}'
@@ -364,7 +364,7 @@ def _shift_amount(helper: str, kwargs: dict[str, int]) -> tuple[str, int]:
     return dim, int(n)
 
 
-def _helper_shift(array: Any, **kwargs: int) -> Any:
+def _helper_shift(array: Any, **kwargs: float) -> Any:
     """Non-cyclic shift along a dimension; vacated positions contribute zero.
 
     Usage in YAML: ``shift(soc, snapshot=1)`` — the value at *t-1*, with the
@@ -379,7 +379,7 @@ def _helper_shift(array: Any, **kwargs: int) -> Any:
     raise TypeError(msg)
 
 
-def _helper_roll(array: Any, **kwargs: int) -> Any:
+def _helper_roll(array: Any, **kwargs: float) -> Any:
     """Roll (circular shift) *array* along a dimension.
 
     Usage in YAML: ``roll(soc, snapshot=1)``
